requirements.txt
PyQt5>=5.15.0
cryptography>=3.4.0
pycryptodome>=3.15.0